Began adding enemy attack logic

+++ Battle-Jawn/src/main/java/com/battlejawn/EnemyMove/Bite.java
@@ -14,7 +14,7 @@ public class Bite implements CriticalHit, Missable {
         if (miss()) {
             setDamage(0);
         } else if (criticalHit()){
-            setDamage(damage *= 1.5);
+            setDamage(damage *= (int) 1.5);
         }
     }
